Fix some Sonar bugs (#821)
* Fix bug in CodeRepositoryProviderHelper

* Fix bug in IncrementalCompiler_New

* Fix bug in CompiledSupport

* Fix bug in TestTDS

+++ legend-pure-functions/legend-pure-functions-relation/legend-pure-runtime-java-extension-shared-functions-relation/src/main/java/org/finos/legend/pure/runtime/java/extension/external/relation/shared/TestTDS.java
@@ -577,7 +577,7 @@ public class TestTDS
                     throw new RuntimeException("ERROR " + copy.columnType.get(columnName) + " not supported in slice!");
             }
         });
-        copy.rowCount = to - from;
+        copy.rowCount = (long) to - from;
         return copy;
     }
